# liveramp-clean-room-api-specification.yml
openapi: 3.0.0
info:
  version: Generic
  title: External APIs for Customer Integration
  description: >-
    APIs to create and fetch Cleanroom and related details from Habu for Customer Integration.
  contact:
    name: Habu
    url: https://habu.com/
    email: platform@habu.com
security:
  - application: [ ]
paths:
  /cleanrooms:
    post:
      summary: Create Cleanroom
      operationId: createCleanroom
      description: >-
        This operation creates a new Cleanroom.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CleanRoomRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/CleanRoomRequest'
      responses:
        '200':
          description: Cleanroom created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CleanRoom"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/ResourceConflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Cleanrooms
    get:
      summary: Fetch a list of all Cleanrooms
      operationId: getAllCleanrooms
      description: >-
        This operation fetches a list of all the Cleanrooms that are created for the Customer Integration.
      parameters:
        - name: limit
          in: query
          description: The number of Cleanrooms returned per page - default is 500
          schema:
            type: integer
            default: 500
        - name: offset
          in: query
          description: The offset - default is 0
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successfully fetched Cleanroom details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CleanRooms'
            application/xml:
              schema:
                $ref: '#/components/schemas/CleanRooms'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Cleanrooms
...
  /cleanrooms/{cleanroomId}/cleanroom-questions:
    post:
      summary: Add Cleanroom Question
      operationId: addCleanroomQuestion
      description: >-
        This operation adds a question to an existing cleanroom.
      parameters:
        - in: path
          name: cleanroomId
          description: Cleanroom ID associated with the Cleanroom Question
          schema:
            type: string
          required: true
        - in: query
          name: questionId
          description: Question ID
          required: true
          schema:
            type: string
        - in: query
          name: isEditable
          description: Whether the resultant Cleanroom Question can be edited further
          schema:
            type: boolean
            default: true
          required: false
      responses:
        '200':
          description: Cleanroom Question created successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CleanRoomQuestionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Cleanroom Questions

    get:
      summary: Fetch a list of all Cleanroom Questions for a Cleanroom
      operationId: getAllCleanroomQuestions
      description: >-
        This operation fetches a list of all the Cleanroom Questions that are tied to a Cleanroom.
      parameters:
        - name: cleanroomId
          in: path
          description: The cleanroom ID to fetch all the underlying Cleanroom Questions
          schema:
            type: string
          required: true
        - name: startDate
          in: query
          description: A filter applied on the CreatedOn field for a Cleanroom Question that marks as the beginning of the range of dates. Default is empty - meaning no filter applied. The format is yyyy-MM-dd
          schema:
            type: string
            format: date
            example: 2024-01-01
        - name: endDate
          in: query
          description: A filter applied on the CreatedOn field for a Cleanroom Question that marks as the end of the range of dates. Default is empty - meaning no filter applied. The format is yyyy-MM-dd
          schema:
            type: string
            format: date
            example: 2024-01-01
        - name: limit
          in: query
          description: The number of Cleanroom Questions returned per page - default is 500
          schema:
            type: integer
            default: 500
        - name: offset
          in: query
          description: The offset - default is 0
          schema:
            type: integer
            default: 0
        - name: questionType
          in: query
          description: List or Analytical
          schema:
            type: string
      responses:
        '200':
          description: Successfully fetched Cleanroom Question details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CleanRoomQuestions'
            application/xml:
              schema:
                $ref: '#/components/schemas/CleanRoomQuestions'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Cleanroom Questions
...
  /cleanroom-questions/{cleanroomQuestionId}/create-run:
    post:
      summary: Create a Cleanroom Question Run for a Cleanroom Question
      operationId: createCleanroomQuestionRun
      description: >-
        This operation facilitates creation of a Cleanroom Question Run for a given Cleanroom Question.
      parameters:
        - name: cleanroomQuestionId
          in: path
          description: The Cleanroom Question ID for which the Cleanroom Question Run is being created
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionRunCreateParameters'
          application/xml:
            schema:
              $ref: '#/components/schemas/QuestionRunCreateParameters'
      responses:
        '200':
          description: Successfully created Cleanroom Question Run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionRunMetadata'
            application/xml:
              schema:
                $ref: '#/components/schemas/QuestionRunMetadata'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Cleanroom Question Runs
  /cleanroom-question-runs/{cleanroomQuestionRunId}:
    get:
      summary: Fetch a Cleanroom Question Run by ID
      operationId: getCleanroomQuestionRunById
      description: >-
        This operation fetches a Cleanroom Question Run (including metadata) by Cleanroom Question Run ID.
      parameters:
        - name: cleanroomQuestionRunId
          in: path
          description: The Cleanroom Question Run ID to fetch a single Cleanroom Question Run's details
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successfully fetched Cleanroom Question Run details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionRunMetadata'
            application/xml:
              schema:
                $ref: '#/components/schemas/QuestionRunMetadata'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Cleanroom Question Runs

    put:
      summary: Update Question Run Status
      operationId: updateCleanroomQuestionRunStatus
      description: >-
        This operation lets the Customer share an update on the status of an in-progress Cleanroom Question Run with Habu. It also lets them send the status and details for a completed/failed Cleanroom Question Run to Habu.
      parameters:
        - name: cleanroomQuestionRunId
          in: path
          description: The Cleanroom Question Run ID for which the update request is being sent
          schema:
            type: string
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionRunStatus'
          application/xml:
            schema:
              $ref: '#/components/schemas/QuestionRunStatus'
      responses:
        '200':
          $ref: '#/components/responses/QuestionRunStatusUpdate'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Cleanroom Question Runs

    delete:
      summary: Delete Cleanroom Question Run
      operationId: deleteCleanroomQuestionRun
      description: >-
        This operation lets the Customer delete a Cleanroom Question Run by ID.
      parameters:
        - name: cleanroomQuestionRunId
          in: path
          description: The Cleanroom Question Run ID
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Cleanroom Question Run deleted successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Cleanroom Question Runs
  /cleanroom-question-runs/{cleanroomQuestionRunId}/data:
    get:
      summary: Get the result data for Cleanroom Question Run
      operationId: getCleanroomQuestionRunData
      description: >-
        This operation is to get the result data for the Cleanroom Question Run.
      parameters:
        - name: cleanroomQuestionRunId
          in: path
          description: The Cleanroom Question Run ID corresponding to this run
          schema:
            type: string
          required: true
        - name: limit
          in: query
          description: The number of Cleanroom Question Run results data returned per page - default is 500
          schema:
            type: integer
            default: 500
        - name: offset
          in: query
          description: The offset - default is 0
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Successfully fetched Cleanroom Question Run Result Data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionRunResultData'
            application/xml:
              schema:
                $ref: '#/components/schemas/QuestionRunResultData'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
      tags:
        - Cleanroom Question Runs
...
components:
  responses:
    BadRequest:
      description: Bad Request - Incorrect syntax or request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReturnObject'
        application/xml:
          schema:
            $ref: '#/components/schemas/ReturnObject'
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReturnObject'
        application/xml:
          schema:
            $ref: '#/components/schemas/ReturnObject'
    Unauthorized:
      description: Authorization information was missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReturnObject'
        application/xml:
          schema:
            $ref: '#/components/schemas/ReturnObject'
    Forbidden:
      description: This action is not allowed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReturnObject'
        application/xml:
          schema:
            $ref: '#/components/schemas/ReturnObject'
    ResourceConflict:
      description: Conflict with the current state of the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReturnObject'
        application/xml:
          schema:
            $ref: '#/components/schemas/ReturnObject'
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ReturnObject'
        application/xml:
          schema:
            $ref: '#/components/schemas/ReturnObject'
...
  schemas:
...
    CleanRoom:
      description: Model containing the Cleanroom details
      properties:
        id:
          type: string
        displayId:
          type: string
        name:
          type: string
        description:
          type: string
        ownerOrganization:
          type: string
        partners:
          type: array
          items:
            type: string
        startAt:
          type: string
        endAt:
          type: string
        questionsCount:
          type: integer
        status:
          type: string
        cleanRoomParameters:
          type: object
          additionalProperties:
            type: string
        timeAudit:
          $ref: "#/components/schemas/TimeAudit"
        cleanRoomTypeId:
          type: string
...
    CleanRooms:
      description: A list of Cleanrooms and their details
      type: array
      items:
        $ref: '#/components/schemas/CleanRoom'
    CleanRoomQuestions:
      description: A list of Cleanroom Questions and their details
      type: array
      items:
        $ref: '#/components/schemas/CleanRoomQuestion'
...
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
    CleanRoomRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the Cleanroom
        description:
          type: string
          description: Description of the Cleanroom
        startAt:
          type: string
          description: Start Date for the cleanroom, use format yyyy-mm-dd
        endAt:
          type: string
          description: End Date for the cleanroom, use format yyyy-mm-dd
        type:
          $ref: "#/components/schemas/CleanRoomTypeRequest"
        parameters:
          $ref: "#/components/schemas/CleanRoomParameterRequest"

      required:
        - organizationId
        - name
        - startAt
        - type
...
    CleanRoomQuestionResponse:
      type: object
      properties:
        id:
          type: string
          description: Cleanroom question ID
        cleanRoomId:
          type: string
          description: Cleanroom ID
        ownerOrganizationId:
          type: string
          description: OrganizationID that brought the question into the clean room
        questionId:
          type: string
          description: Question ID
        title:
          type: string
          description: Title of the Question
        category:
          type: string
          description: Category of the Question
        stage:
          type: string
          description: Stage of the Cleanroom Question
        status:
          type: string
        latestRun:
          $ref: '#/components/schemas/CleanRoomQuestionRun'
        datasetCount:
          type: integer
          description: Count of required datasets for this Question
        runCount:
          type: integer
          description: Count of Report Runs
        isEditable:
          type: boolean
          description: Is Editable
        questionStatus:
          type: string
        questionVersion:
          type: integer
          description: Question Version
        displayId:
          type: string
          description: display ID
        description:
          type: string
          description: Description of the Question
...
  securitySchemes:
    application:
      type: oauth2
      flows:
        clientCredentials:
          tokenUrl: 'https://api.habu.com/v1/oauth/token'
          scopes: { }
servers:
  - description: External APIs for Customer Integration
    url: https://api.habu.com/v1/